# src/main/resources/openapi/api-spec.yaml
openapi: 3.0.3
info:
  title: Captain API
  description: API for Captain application generated from OpenAPI specification
  version: 1.0.0
  contact:
    name: Captain Team
    email: support@captain.com

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Users
    description: User management operations
  - name: Items
    description: Item management operations

paths:
  /api/users:
    get:
      tags:
        - Users
      summary: Get all users
      description: Retrieve a list of all users
      operationId: getAllUsers
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags:
        - Users
      summary: Create a new user
      description: Create a new user in the system
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/users/{userId}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Retrieve a specific user by their ID
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          required: true
          description: ID of the user to retrieve
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Users
      summary: Update user
      description: Update an existing user
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          description: ID of the user to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Users
      summary: Delete user
      description: Delete a user by ID
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          description: ID of the user to delete
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/items:
    get:
      tags:
        - Items
      summary: Get all items
      description: Retrieve a list of all items
      operationId: getAllItems
      parameters:
        - name: category
          in: query
          required: false
          description: Filter items by category
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Items
      summary: Create a new item
      description: Create a new item in the system
      operationId: createItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemRequest'
      responses:
        '201':
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    User:
      type: object
      required:
        - id
        - username
        - email
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the user
        username:
          type: string
          description: Username of the user
          example: johndoe
        email:
          type: string
          format: email
          description: Email address of the user
          example: john.doe@example.com
        firstName:
          type: string
          description: First name of the user
          example: John
        lastName:
          type: string
          description: Last name of the user
          example: Doe
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the user was created

    UserRequest:
      type: object
      required:
        - username
        - email
      properties:
        username:
          type: string
          description: Username of the user
          minLength: 3
          maxLength: 50
          example: johndoe
        email:
          type: string
          format: email
          description: Email address of the user
          example: john.doe@example.com
        firstName:
          type: string
          description: First name of the user
          example: John
        lastName:
          type: string
          description: Last name of the user
          example: Doe

    Item:
      type: object
      required:
        - id
        - name
        - category
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the item
        name:
          type: string
          description: Name of the item
          example: Laptop
        description:
          type: string
          description: Description of the item
          example: High-performance laptop for development
        category:
          type: string
          description: Category of the item
          example: Electronics
        price:
          type: number
          format: double
          description: Price of the item
          example: 999.99
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the item was created

    ItemRequest:
      type: object
      required:
        - name
        - category
      properties:
        name:
          type: string
          description: Name of the item
          minLength: 1
          maxLength: 100
          example: Laptop
        description:
          type: string
          description: Description of the item
          example: High-performance laptop for development
        category:
          type: string
          description: Category of the item
          example: Electronics
        price:
          type: number
          format: double
          description: Price of the item
          minimum: 0
          example: 999.99

    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred